tech(e2e): extract browser contexts fixture
