+++ high-level-tests/e2e-playwright/fixtures/browser-contexts.ts
@@ -2,7 +2,6 @@ import path from 'node:path';
 
 import type { BrowserContext } from '@playwright/test';
 import { test as base } from '@playwright/test';
-import crypto from 'crypto';
 
 import {
   AUTH_DIR,
@@ -16,7 +15,7 @@ import {
 const shouldRecordHAR = process.env.RECORD_HAR === 'true';
 const HAR_DIR = path.resolve(import.meta.dirname, '../.har-record');
 
-export const sharedFixtures = base.extend<{
+export const browserContextsFixtures = base.extend<{
   globalTestId: string;
   forEachTest: void;
   pixAppUserContext: BrowserContext;
@@ -25,32 +24,6 @@ export const sharedFixtures = base.extend<{
   pixCertifProContext: BrowserContext;
   pixSuperAdminContext: BrowserContext;
 }>({
-  // eslint-disable-next-line no-empty-pattern
-  globalTestId: async ({}, use, testInfo) => {
-    const raw = `${testInfo.file}::${testInfo.title}`;
-    const hash = crypto.createHash('sha1').update(raw).digest('hex');
-    await use(hash);
-  },
-  page: async ({ context }, use) => {
-    const page = await context.newPage();
-    await use(page);
-    await page.close();
-  },
-  context: async ({ browser }, use, testInfo) => {
-    const harFilePath = path.join(HAR_DIR, `${sanitizeFilename(testInfo.title)}-defaultContext.har`);
-    const context = await browser.newContext(
-      shouldRecordHAR
-        ? {
-            recordHar: {
-              path: harFilePath,
-              content: 'omit',
-            },
-          }
-        : {},
-    );
-    await use(context);
-    await context.close();
-  },
   pixAppUserContext: async ({ browser }, use, testInfo) => {
     const authFilePath = path.join(AUTH_DIR, `${PIX_APP_USER_CREDENTIALS.label}.json`);
     const harFilePath = path.join(HAR_DIR, `${sanitizeFilename(testInfo.title)}-${PIX_APP_USER_CREDENTIALS.label}.har`);
