+++ high-level-tests/e2e-playwright/fixtures/index.ts
@@ -1,8 +1,9 @@
 import { mergeTests } from '@playwright/test';
 
-import { sharedFixtures } from './shared.ts';
+import { browserContextsFixtures } from './browser-contexts.ts';
 import { snapshotFixtures } from './snapshot.ts';
+import { utilsFixtures } from './utils.ts';
 
-export const test = mergeTests(sharedFixtures, snapshotFixtures);
+export const test = mergeTests(utilsFixtures, snapshotFixtures, browserContextsFixtures);
 
 export const expect = test.expect;
